+++ src/routes/InstancesRoute.js
@@ -1,50 +1,20 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import {
-  omit,
   keyBy,
   get,
-  set,
   template,
-  flowRight,
 } from 'lodash';
-import {
-  injectIntl,
-  intlShape,
-} from 'react-intl';
-import {
-  stripesShape,
-  AppIcon,
-  stripesConnect,
-} from '@folio/stripes/core';
-import {
-  SearchAndSort,
-  makeQueryFunction,
-} from '@folio/stripes/smart-components';
+import { stripesConnect } from '@folio/stripes/core';
+import { makeQueryFunction } from '@folio/stripes/smart-components';
 
-import {
-  FilterNavigation,
-  InstanceFilters,
-  HoldingFilters,
-  ItemFilters,
-} from '../components';
-import packageInfo from '../../package';
-import InstanceForm from '../edit/InstanceForm';
-import ViewInstance from '../ViewInstance';
-import formatters from '../referenceFormatters';
-import withLocation from '../withLocation';
-import {
-  getCurrentFilters,
-  parseFiltersToStr,
-  getSegment,
-} from '../utils';
+import InstancesView from '../views';
 import {
   searchableIndexes,
   filterConfig,
 } from '../constants';
 
 const INITIAL_RESULT_COUNT = 30;
-const RESULT_COUNT_INCREMENT = 30;
 
 class InstancesRoute extends React.Component {
   static defaultProps = {
@@ -228,51 +198,6 @@ class InstancesRoute extends React.Component {
     },
   });
 
-  constructor(props) {
-    super(props);
-
-    this.cViewInstance = this.props.stripes.connect(ViewInstance);
-    this.resultsList = null;
-    this.SRStatus = null;
-    this.copyInstance = this.copyInstance.bind(this);
-
-    this.state = {};
-    this.filterRenderers = {
-      instances: this.renderInstanceFilters,
-      holdings: this.renderHoldingFilters,
-      items: this.renderItemFilters,
-    };
-  }
-
-  onChangeIndex = (e) => {
-    const qindex = e.target.value;
-    this.props.updateLocation({ qindex });
-  };
-
-  onFilterChangeHandler = ({ name, values }) => {
-    const { resources: { query } } = this.props;
-    const curFilters = getCurrentFilters(get(query.filters));
-    const mergedFilters = values.length
-      ? { ...curFilters, [name]: values }
-      : omit(curFilters, name);
-    const filtersStr = parseFiltersToStr(mergedFilters);
-
-    this.props.updateLocation({ filters: filtersStr });
-  };
-
-  closeNewInstance = (e) => {
-    if (e) e.preventDefault();
-    this.setState({ copiedInstance: null });
-    this.props.updateLocation({ layer: null });
-  };
-
-  copyInstance(instance) {
-    let copiedInstance = omit(instance, ['id', 'hrid']);
-    copiedInstance = set(copiedInstance, 'source', 'FOLIO');
-    this.setState({ copiedInstance });
-    this.props.updateLocation({ layer: 'create' });
-  }
-
   createInstance = (instance) => {
     // POST item record
     this.props.mutator.records.POST(instance).then(() => {
@@ -294,75 +219,22 @@ class InstancesRoute extends React.Component {
     return false;
   }
 
-  getReferenceTables() {
+  getData() {
     const { manifest } = InstancesRoute;
     const { resources } = this.props;
-    const referenceTables = {};
+    const data = {};
 
     for (const key in manifest) {
       if (key !== 'records' && manifest[key].type === 'okapi') {
-        referenceTables[key] = get(resources, `${key}.records`, []);
+        data[key] = get(resources, `${key}.records`, []);
       }
     }
 
-    referenceTables.locationsById = keyBy(referenceTables.locations, 'id');
+    data.locationsById = keyBy(data.locations, 'id');
+    data.query = resources.query;
 
-    return referenceTables;
+    return data;
   }
-
-  renderNavigation = () => (
-    <FilterNavigation segment={getSegment(this.props.getParams())} />
-  );
-
-  renderInstanceFilters = (activeFilters, onChange) => {
-    const { resources } = this.props;
-
-    return (
-      <InstanceFilters
-        activeFilters={activeFilters}
-        data={{
-          locations: get(resources, 'locations.records', []),
-          resourceTypes: get(resources, 'instanceTypes.records', []),
-        }}
-        onChange={onChange}
-        onClear={(name) => onChange({ name, values: [] })}
-      />
-    );
-  }
-
-  renderHoldingFilters = (activeFilters, onChange) => {
-    return (
-      <HoldingFilters
-        activeFilters={activeFilters}
-        data={{
-          // TODO: pass required data
-        }}
-        onChange={onChange}
-        onClear={(name) => onChange({ name, values: [] })}
-      />
-    );
-  }
-
-  renderItemFilters = (activeFilters, onChange) => {
-    return (
-      <ItemFilters
-        activeFilters={activeFilters}
-        data={{
-          // TODO: pass required data
-        }}
-        onChange={onChange}
-        onClear={(name) => onChange({ name, values: [] })}
-      />
-    );
-  }
-
-  renderFilters = (onChange) => {
-    const { resources, getParams } = this.props;
-    const segment = getSegment(getParams());
-    const activeFilters = getCurrentFilters(get(resources, 'query.filters', ''));
-
-    return this.filterRenderers[segment](activeFilters, onChange);
-  };
 
   render() {
     const {
@@ -370,89 +242,30 @@ class InstancesRoute extends React.Component {
       browseOnly,
       onSelectRow,
       disableRecordCreation,
-      visibleColumns,
-      intl,
+      resources,
+      mutator,
     } = this.props;
 
     if (this.isLoading()) {
       return null;
     }
 
-    const referenceTables = this.getReferenceTables();
-    const resultsFormatter = {
-      'title': ({ title }) => (
-        <AppIcon
-          size="small"
-          app="inventory"
-          iconKey="instance"
-          iconAlignment="baseline"
-        >
-          {title}
-        </AppIcon>
-      ),
-      'relation': r => formatters.relationsFormatter(r, referenceTables.instanceRelationshipTypes),
-      'publishers': r => r.publication.map(p => (p ? `${p.publisher} ${p.dateOfPublication ? `(${p.dateOfPublication})` : ''}` : '')).join(', '),
-      'publication date': r => r.publication.map(p => p.dateOfPublication).join(', '),
-      'contributors': r => formatters.contributorsFormatter(r, referenceTables.contributorTypes),
-    };
-
-    const formattedSearchableIndexes = searchableIndexes.map(index => {
-      const { prefix = '' } = index;
-      const label = prefix + intl.formatMessage({ id: index.label });
-
-      return { ...index, label };
-    });
-
     return (
-      <div data-test-inventory-instances>
-        <SearchAndSort
-          packageInfo={packageInfo}
-          objectName="inventory"
-          maxSortKeys={1}
-          renderNavigation={this.renderNavigation}
-          searchableIndexes={formattedSearchableIndexes}
-          selectedIndex={get(this.props.resources.query, 'qindex')}
-          searchableIndexesPlaceholder={null}
-          onChangeIndex={this.onChangeIndex}
-          initialResultCount={INITIAL_RESULT_COUNT}
-          resultCountIncrement={RESULT_COUNT_INCREMENT}
-          viewRecordComponent={ViewInstance}
-          editRecordComponent={InstanceForm}
-          newRecordInitialValues={(this.state && this.state.copiedInstance) ? this.state.copiedInstance : {
-            discoverySuppress: false,
-            staffSuppress: false,
-            previouslyHeld: false,
-            source: 'FOLIO'
-          }}
-          visibleColumns={visibleColumns || ['title', 'contributors', 'publishers', 'relation']}
-          columnMapping={{
-            title: intl.formatMessage({ id: 'ui-inventory.instances.columns.title' }),
-            contributors: intl.formatMessage({ id: 'ui-inventory.instances.columns.contributors' }),
-            publishers: intl.formatMessage({ id: 'ui-inventory.instances.columns.publishers' }),
-            relation: intl.formatMessage({ id: 'ui-inventory.instances.columns.relation' }),
-          }}
-          columnWidths={{ title: '40%' }}
-          resultsFormatter={resultsFormatter}
-          onCreate={this.createInstance}
-          viewRecordPerms="ui-inventory.instance.view"
-          newRecordPerms="ui-inventory.instance.create"
-          disableRecordCreation={disableRecordCreation || false}
-          parentResources={this.props.resources}
-          parentMutator={this.props.mutator}
-          detailProps={{ referenceTables, onCopy: this.copyInstance }}
-          path={`${this.props.match.path}/(view|viewsource)/:id/:holdingsrecordid?/:itemid?`}
-          showSingleResult={showSingleResult}
-          browseOnly={browseOnly}
-          onSelectRow={onSelectRow}
-          renderFilters={this.renderFilters}
-          onFilterChange={this.onFilterChangeHandler}
-        />
-      </div>);
+      <InstancesView
+        parentResources={resources}
+        parentMutator={mutator}
+        data={this.getData()}
+        browseOnly={browseOnly}
+        showSingleResult={showSingleResult}
+        onCreate={this.createInstance}
+        onSelectRow={onSelectRow}
+        disableRecordCreation={disableRecordCreation}
+      />
+    );
   }
 }
 
 InstancesRoute.propTypes = {
-  stripes: stripesShape.isRequired,
   resources: PropTypes.shape({
     records: PropTypes.shape({
       hasLoaded: PropTypes.bool.isRequired,
@@ -509,14 +322,6 @@ InstancesRoute.propTypes = {
   browseOnly: PropTypes.bool,
   disableRecordCreation: PropTypes.bool,
   onSelectRow: PropTypes.func,
-  visibleColumns: PropTypes.arrayOf(PropTypes.string),
-  updateLocation: PropTypes.func.isRequired,
-  getParams: PropTypes.func.isRequired,
-  intl: intlShape,
 };
 
-export default flowRight(
-  stripesConnect,
-  injectIntl,
-  withLocation,
-)(InstancesRoute);
+export default stripesConnect(InstancesRoute);
